Fix device and remove std and mean in tokenizer
