+++ examples/rt1_test.py
@@ -1,22 +1,34 @@
+import numpy as np
 import torch
-from mbodied_agents.agents.models.rt1.rt1_module import RT1Module
+from gym import spaces
+from mbodied_agents.agents.motion.rt1.rt1_agent import RT1Agent
+
+# Define the observation and action spaces
+observation_space = spaces.Dict({
+    'image': spaces.Box(low=0, high=255, shape=(224, 224, 3), dtype=np.uint8),
+    'instruction': spaces.Discrete(10)
+})
+action_space = spaces.Dict({
+    'gripper_position': spaces.Box(low=-1, high=1, shape=(3,), dtype=np.float32),
+    'gripper_action': spaces.Discrete(2)
+})
 
 config = {
-    'observation_history_size': 10,
-    'future_action_window_size': 10,
-    'token_embedding_dim': 512,
-    'causal_attention': True,
-    'num_layers': 4,
-    'layer_size': 256,
+    "num_layers": 8,
+    "layer_size": 128,
+    "observation_history_size": 6,
+    "future_prediction": 6,
+    "token_embedding_dim": 512,
+    "causal_attention": True,
 }
 
-# model = RT1Module(config)
+rt1_agent = RT1Agent(config)
 
-# Create a dummy input
-dummy_input = {
-    'image_primary': torch.rand(1, 3, 224, 224),  # Batch size of 1
-    'natural_language_embedding': torch.rand(1, 512),
-}
+for name, param in rt1_agent.model.named_parameters():
+    print(f"{name}: {param.device}")
+
+rt1_agent.act(image=torch.rand(224, 224, 3),
+              instruction_emb=torch.rand(1, 512))
 
 # Pass the dummy input through the model
 # output = model(dummy_input)
