+++ src/mbodied_agents/base/backend.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Mbodi AI
+˛# Copyright 2024 Mbodi AI
 # 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
