+++ src/mbodied_agents/base/motion.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 import logging
+
 import numpy as np
+from mbodied_agents.base.sample import Sample
+from mbodied_agents.types.ndarray import NumpyArray
 from pydantic import ConfigDict, Field
 from typing_extensions import Literal, get_origin
-
-from mbodied_agents.types.ndarray import NumpyArray
-from mbodied_agents.base.sample import Sample
 
 MotionSubtype = Literal[
     "unspecified",
