+++ src/mbodied_agents/types/vision.py
@@ -23,13 +23,12 @@ from urllib.request import urlopen
 
 import numpy as np
 from gym import spaces
+from mbodied_agents.base.sample import Sample
+from mbodied_agents.types.ndarray import NumpyArray
 from PIL import Image as PILModule
 from PIL.Image import Image as PILImage
 from pydantic import AnyUrl, Base64Str, ConfigDict, Field, FilePath, InstanceOf, field_serializer, model_validator
 from typing_extensions import Literal
-
-from mbodied_agents.base.sample import Sample
-from mbodied_agents.types.ndarray import NumpyArray
 
 SupportsImage = Union[np.ndarray, PILImage, Base64Str, AnyUrl, FilePath]
